Revert "Merge branch 'main' of https://github.com/TienLee2/PRU-Project"
This reverts commit a25549b0481b05e5624c84143775ac689b63b12f.

+++ Assets/HealthBar.cs
@@ -9,15 +9,13 @@ public class HealthBar : MonoBehaviour
     public Slider slider;
     public Gradient gradient;
     public Image fill;
-
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
         slider.value = health;
-        fill.color = gradient.Evaluate(1f);
+      fill.color = gradient.Evaluate(1f);
     }
-
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
